Renamed and mark ContextAware like ContextAwareUtils.
